fixup! test: add tests for TaskRunV2 components and kubearchive integration
